requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
duckdb==1.1.2
Faker==19.13.0
Flask==3.0.3
fonttools==4.54.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain-core==0.3.12
langchain-groq==0.2.0
langsmith==0.1.137
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==1.23.5
openai==1.52.2
orjson==3.10.10
packaging==24.1
pandas==1.5.3
pandasai==2.3.0
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlglot==25.27.0
sqlglotrs==0.2.12
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
